extract modules to make code structure more apparent
